requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aom=3.4.0=hbe31e94_1
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
brotlipy=0.7.0=py39hb18efdd_1004
bzip2=1.0.8=h3422bc3_4
ca-certificates=2022.6.15=h4653dfc_0
certifi=2022.6.15=py39h2804cbe_0
cffi=1.15.1=py39h04d3946_0
charset-normalizer=2.1.0=pyhd8ed1ab_0
cloudpickle=2.1.0=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
cryptography=37.0.1=py39h834c97f_0
cycler=0.11.0=pyhd8ed1ab_0
dgl=0.9.0=pypi_0
expat=2.4.8=h6b3803e_0
ffmpeg=5.1.0=gpl_hf4fd261_101
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=hfb34624_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.35.0=py39h4eb3d34_0
freetype=2.12.1=hd633e50_0
gettext=0.19.8.1=h049c9fb_1008
gmp=6.2.1=h9f76cd9_0
gnutls=3.7.6=h27a1928_5
gym=0.21.0=py39h2804cbe_2
icu=70.1=h6b3803e_0
idna=3.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9e=he4db4b2_2
kiwisolver=1.4.4=py39hab5e169_0
lame=3.100=h27ca646_1001
lcms2=2.12=had6a04f_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=16_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=16_osxarm64_openblas
libcxx=14.0.6=h04bba0f_0
libdeflate=1.13=h57fd34a_0
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0.dev0=11_0_1_hf114ba7_23
libgfortran5=11.0.1.dev0=hf114ba7_23
libiconv=1.16=h642e427_0
libidn2=2.3.3=he4db4b2_0
liblapack=3.9.0=16_osxarm64_openblas
libopenblas=0.3.21=openmp_hcb59c3b_0
libpng=1.6.37=h9b2ccdc_4
libsqlite=3.39.2=h2c9beb0_1
libtasn1=4.18.0=he4db4b2_1
libtiff=4.4.0=hd6ca8d7_3
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.9.14=h9d8dfc2_4
libzlib=1.2.12=ha287fd2_2
llvm-openmp=14.0.4=hd125106_0
lz4-c=1.9.3=hbdafb3b_1
matplotlib-base=3.5.3=py39h267d6a8_0
mpi=1.0=mpich
mpi4py=3.1.3=py39h65280c7_2
mpich=4.0.2=hd4b5bf3_100
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h07bb92c_1
nettle=3.8.1=h63371fa_1
networkx=2.8.5=pypi_0
numpy=1.22.4=py39h7df2422_0
openh264=2.3.0=h9a09cb3_0
openjpeg=2.5.0=h5d4e404_1
openssl=3.0.5=h7aea29f_1
p11-kit=0.24.1=h29577a5_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.3=py39hd2dba81_0
patsy=0.5.2=pyhd8ed1ab_0
pillow=9.2.0=py39he45c975_2
pip=22.2.2=pyhd8ed1ab_0
psutil=5.9.1=pypi_0
pthread-stubs=0.4=h27ca646_1001
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py39h2804cbe_5
python=3.9.13=h96fcbfb_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_0
pytz=2022.2.1=pyhd8ed1ab_0
pyyaml=6.0=py39hb18efdd_4
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_0
scipy=1.9.0=py39h14896cb_0
seaborn=0.11.2=hd8ed1ab_0
seaborn-base=0.11.2=pyhd8ed1ab_0
setuptools=65.0.1=py39h2804cbe_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.39.2=h40dfcc0_1
statsmodels=0.13.2=py39h7b9fbcb_0
svt-av1=1.2.0=he23bcac_0
tk=8.6.12=he1e0b03_0
torchaudio=0.12.1=py39_cpu
torchvision=0.13.1=py39_cpu
tqdm=4.64.0=pyhd8ed1ab_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022c=h191b570_0
unicodedata2=14.0.0=py39hb18efdd_1
urllib3=1.26.11=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zstd=1.5.2=h68b4297_4
